+++ packages/cmf-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.5.6
+
+### Patch Changes
+
+- 667cd0a50: chore: upgrade dependencies and align @talend scoped packages to latest
+- Updated dependencies [667cd0a50]
+  - @talend/react-cmf@6.38.4
+
 ## 3.5.5
 
 ### Patch Changes