+++ packages/containers/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @talend/react-containers
 
+## 6.39.0
+
+### Minor Changes
+
+- 41f8343e2: feat: add support for initial state
+
+### Patch Changes
+
+- Updated dependencies [d5f261f49]
+  - @talend/react-components@6.41.1
+  - @talend/react-forms@6.41.3
+
+## 6.38.0
+
+### Minor Changes
+
+- b55f6de12: feat(router-bridge/SidePanel): support route basename
+
+### Patch Changes
+
+- 667cd0a50: chore: upgrade dependencies and align @talend scoped packages to latest
+- Updated dependencies [667cd0a50]
+- Updated dependencies [2d9e62a2a]
+- Updated dependencies [7dde61e46]
+  - @talend/react-cmf@6.38.4
+  - @talend/react-components@6.40.0
+  - @talend/react-forms@6.41.0
+
 ## 6.37.3
 
 ### Patch Changes